Generate fresh Message-Id headers to work around Gmail's cleverness
When APPENDing to Gmail's IMAP, if the Message-Id of the uploaded
message has alrady been seen before, it assumes it's the same message
and ignores the content of the appended message. This breaks the tests
when the test message is changed.

A new Message-Id is now generated each time to work around this.